requirements.txt
Django>=5.1.0,<5.2.0
djangorestframework>=3.14.0,!=3.15.0
psycopg2>=2.9.10,<3.0
drf-spectacular>=0.28.0,<0.29.0
Pillow
